requirements.txt
# Django and DRF
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Database
psycopg2-binary==2.9.10

# WebSockets (Channels)
channels==4.2.0
channels-redis==4.2.1

# Celery
celery==5.4.0
django-celery-beat==2.8.0
django-celery-results==2.5.1
redis==5.2.1

# Stripe
stripe==11.3.0

# Email & Other
Pillow==11.0.0
python-dotenv==1.0.1
